+++ docs/changelog.md
@@ -35,6 +35,8 @@
   - added `IsBoolean::check()`
   - added `is_boolean()` convenience function
   - added `is_boolean_list()` convenience function
+  - added `IsCallable::check()`
+  - added `is_callable_list()` convenience function
 * Added type-hinting / strict type declaration support for entities:
   - added empty `Entity` interface
 * Added immutable support to entities